+++ test/uploaded_file_test.rb
@@ -319,22 +319,6 @@ describe Shrine::UploadedFile do
       assert_match "file", uploaded_file.download.read
     end
 
-    it "deletes the Tempfile if there's an error in downloading the file (Shrine class level)" do
-      @uploader.storage.instance_eval { undef download }
-      Tempfile.stub(:new, tempfile = Tempfile.new("foo")) do
-        assert_raises(KeyError) { uploaded_file.download }
-      end
-      assert tempfile.closed?
-      assert_nil tempfile.path
-    end
-
-    it "deletes tempfile if tempfile exists and there's error in downloading the file" do
-      @uploader.storage.instance_eval { undef download }
-      Tempfile.stub(:new, nil) do
-        assert_raises(KeyError) { uploaded_file.download }
-      end
-    end
-
     it "uses extension from #id" do
       @uploader.storage.instance_eval { undef download }
       uploaded_file = @uploader.upload(fakeio, location: "foo.jpg")
@@ -358,6 +342,21 @@ describe Shrine::UploadedFile do
       @uploader.storage.instance_eval { undef download }
       uploaded_file.expects(:open).with(foo: "bar")
       uploaded_file.download(foo: "bar")
+    end
+
+    it "deletes the Tempfile if an error occurs while retrieving file contents" do
+      @uploader.storage.instance_eval { undef download }
+      tempfile = Tempfile.new("")
+      Tempfile.stubs(:new).returns(tempfile)
+      assert_raises(KeyError) { uploaded_file.download }
+      assert tempfile.closed?
+      assert_nil tempfile.path
+    end
+
+    it "propagates failures in creating tempfiles" do
+      @uploader.storage.instance_eval { undef download }
+      Tempfile.stubs(:new).raises(Errno::EMFILE) # too many open files
+      assert_raises(Errno::EMFILE) { uploaded_file.download }
     end
   end